requirements/base.txt
# Bleeding edge Django
django>=1.7.7,<1.8

# Configuration
django-configurations==0.8
django-secure==1.0.1
django-cache-url==0.8.0
dj-database-url==0.3.0

# Forms
django-braces==1.4.0
django-crispy-forms==1.4.0
django-floppyforms==1.3.0

# Models
django-model-utils==2.2

# Images
Pillow==2.8.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.19.1

# For the persistence stores
MySQL-python==1.2.5

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.7.2

# Useful things
django-avatar==2.0

# Time zones support
pytz==2015.2

# Your custom requirements go here
django-constance[database]
https://github.com/watchdogpolska/django-one-petition/archive/v0.4.zip
https://github.com/watchdogpolska/django-multi-contact/archive/v0.2.zip
